+++ src/voice/voicecontread36.c
@@ -25,7 +25,7 @@ s32 __osVoiceContRead36(OSMesgQueue* mq, int channel, u16 address, u8* buffer) {
 
         ptr = (u8*)&__osPfsPifRam.ramarray;
 
-        if ((__osContLastCmd != CONT_CMD_READ36_VOICE) || ((u32)__osPfsLastChannel != channel)) {
+        if ((__osContLastCmd != CONT_CMD_READ36_VOICE) || (__osPfsLastChannel != channel)) {
             __osContLastCmd = CONT_CMD_READ36_VOICE;
             __osPfsLastChannel = channel;
